pyproject.toml
[project]
name = "notificationbackend"
version = "0.1.0"
description = ""
authors = [
    {name = "Your Name",email = "you@example.com"}
]

readme = "README.md"

packages = [
    { include = "app_project" },
    { include = "app" },
    { include = "celery_config.py" },
    { include = "manage.py" }
]

requires-python = ">=3.12,<3.15"
dependencies = [
    "django (>=5.2.8,<6.0.0)",
    "pydantic (>=2.12.4,<3.0.0)",
    "pydantic-settings (>=2.12.0,<3.0.0)",
    "psycopg2 (>=2.9.11,<3.0.0)",
    "djangorestframework (>=3.16.1,<4.0.0)",
    "drf-spectacular (>=0.29.0,<0.30.0)",
    "redis (>=7.1.0,<8.0.0)",
    "celery (>=5.5.3,<6.0.0)",
    "colorlog (>=6.10.1,<7.0.0)",
    "python-json-logger (>=4.0.0,<5.0.0)",
    "httpx (>=0.28.1,<0.29.0)",
    "python-telegram-bot (>=22.5,<23.0)",
    "dj-database-url (>=3.0.1,<4.0.0)",
    "whitenoise (>=6.11.0,<7.0.0)",
    "django-environ (>=0.12.0,<0.13.0)",
    "cryptography (>=46.0.3,<47.0.0)",
    "torch (>=2.9.1,<3.0.0)",
    "scikit-learn (>=1.7.2,<2.0.0)"
]

[tool.poetry.dependencies.gunicorn]
version = ">=23.0.0,<24.0.0"
platform = "linux"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
